rename file: report-generator-impl -> report-generator
