+++ src/DetailsView/components/report-export-component.tsx
@@ -4,7 +4,7 @@ import { escape } from 'lodash';
 import { ActionButton } from 'office-ui-fabric-react/lib/Button';
 import * as React from 'react';
 import { ExportResultType } from '../../common/telemetry-events';
-import { ReportGenerator } from '../reports/report-generator-impl';
+import { ReportGenerator } from '../reports/report-generator';
 import { ExportDialog, ExportDialogDeps } from './export-dialog';
 
 export type ReportExportComponentDeps = ExportDialogDeps;
