+++ src/DetailsView/reports/report-generator.ts
