+++ src/tests/unit/tests/DetailsView/components/report-export-component.test.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 import { ExportDialog, ExportDialogDeps } from '../../../../../DetailsView/components/export-dialog';
 import { ReportExportComponent, ReportExportComponentProps } from '../../../../../DetailsView/components/report-export-component';
-import { ReportGenerator } from '../../../../../DetailsView/reports/report-generator-impl';
+import { ReportGenerator } from '../../../../../DetailsView/reports/report-generator';
 
 describe('ReportExportComponentTest', () => {
     let deps: ExportDialogDeps;
