+++ src/tests/unit/tests/DetailsView/reports/report-generator-impl.test.ts
@@ -6,7 +6,7 @@ import { AssessmentStoreData } from '../../../../../common/types/store-data/asse
 import { FeatureFlagStoreData } from '../../../../../common/types/store-data/feature-flag-store-data';
 import { TabStoreData } from '../../../../../common/types/store-data/tab-store-data';
 import { AssessmentReportHtmlGenerator } from '../../../../../DetailsView/reports/assessment-report-html-generator';
-import { ReportGenerator } from '../../../../../DetailsView/reports/report-generator-impl';
+import { ReportGenerator } from '../../../../../DetailsView/reports/report-generator';
 import { ReportHtmlGenerator } from '../../../../../DetailsView/reports/report-html-generator';
 import { ReportNameGenerator } from '../../../../../DetailsView/reports/report-name-generator';
 import { ScanResults } from '../../../../../scanner/iruleresults';
@@ -41,11 +41,7 @@ describe('ReportGeneratorImpl', () => {
             )
             .returns(() => 'returned-data');
 
-        const testObject = new ReportGenerator(
-            nameBuilderMock.object,
-            dataBuilderMock.object,
-            assessmentReportHtmlGeneratorMock.object,
-        );
+        const testObject = new ReportGenerator(nameBuilderMock.object, dataBuilderMock.object, assessmentReportHtmlGeneratorMock.object);
         const actual = testObject.generateFastPassAutomateChecksReport(scanResult, date, title, url, description);
 
         expect(actual).toMatchSnapshot();
@@ -65,11 +61,7 @@ describe('ReportGeneratorImpl', () => {
             .returns(() => 'generated-assessment-html')
             .verifiable(Times.once());
 
-        const testObject = new ReportGenerator(
-            nameBuilderMock.object,
-            dataBuilderMock.object,
-            assessmentReportHtmlGeneratorMock.object,
-        );
+        const testObject = new ReportGenerator(nameBuilderMock.object, dataBuilderMock.object, assessmentReportHtmlGeneratorMock.object);
         const actual = testObject.generateAssessmentReport(
             assessmentStoreData,
             assessmentsProvider,
@@ -88,11 +80,7 @@ describe('ReportGeneratorImpl', () => {
             .returns(() => 'returned-name')
             .verifiable(Times.once());
 
-        const testObject = new ReportGenerator(
-            nameBuilderMock.object,
-            dataBuilderMock.object,
-            assessmentReportHtmlGeneratorMock.object,
-        );
+        const testObject = new ReportGenerator(nameBuilderMock.object, dataBuilderMock.object, assessmentReportHtmlGeneratorMock.object);
         const actual = testObject.generateName('InsightsScan', date, title);
 
         const expected = 'returned-name';
